+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed a 1.5.1.4 regression that caused incorrect rocket trajectories
 - Fixed a 1.4.2.0 regression that caused the recipe of gas turbine to be missing
 - Added some other missing crafting recipes
+- Fixed radiation system not diffusing correctly across non-uniform sections
 ### Changes
 - Added new absorber blocks that use block metadata. Old absorber blocks will be transformed automatically.
 ### New Features